requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
python-dotenv>=1.0.0
google-cloud-aiplatform[adk]>=1.88.0
google-auth>=2.0.0
fit-tool>=0.9.15
pytest>=8.0
pytest-asyncio>=1.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
